resolved stop button issue
Stop button previously would not work when IDE was started under some
circumstances (particularly, when it was started from boot init scripts).  The
issue was a pecularity of the handling of SIGINT by Python.  Python usually
would catch the error as a keyboard error, but not always.

If catching KeyboardException was explicit in the code, then the SIGINT is
always handled.  This appears to be a known issue (or feature?) in Python.  So,
one workable fix would be to have some wrapper code that runs when in the IDE
that handles KeyboardException.  A disadvantage is that someone can write code
that tries to handle KeyboardException themselves, and thereby prevent the IDE
from being stoppable.

A perhaps better solution is simply to kill the offending process with SIGTERM,
and that's what is now implemented.  In addition, we now report to the IDE if
the prorgam completed, or was stopped.

Also, an uneeded mutex in the code was removed.