output window performance - fixes #53
Previous temp fix wasn't good enough (helped only when running remotely).  This
fix pushes all buffer handing into the go server, and out of the javascript,
which couldn't handle it.  Now, when the go server is sending output window
updates frequently, it trucates the buffer to send to be just enough to fit in
the output window.  Even then, it doesn't send them faster than 20Hz (which
looks plenty fast).

The output window size is assumed to be less than 20 lines right now.  This
would break (gracefully) if the window is resized - it would be better to send
the size from the javascript to the server.