Removed run-on-boot support.  Closes #86.
Run-on-boot support was not working.  It may be added back in the future,
however, its a potentially dangerous feature, as it could cause the IDE to be
unresponsive depending on the user code that runs.

It may be wise to add a ~10 second countdown before running the app, so it can
be cancelled.